+++ types/index.d.ts
@@ -43,6 +43,7 @@ declare namespace anchor {
     level?: number | number[];
 
     slugify?(str: string): string;
+    getTokensText?(tokens: Token[]): string;
 
     uniqueSlugStartIndex?: number;
     permalink?: PermalinkGenerator;